feat: Change between pages with React Router
Substitute the usage of App state to show either Books list or Books search
with a React Router and separated components for each page (Root and Search).

When we reload the page, we were seeing the Book List page,
even when on the Search page. With React Router we can still see the same
Search page, relying on the browser URL to show either page.

- Create Root and Search page components
- Set BrowserRouter in src/index and renders App withing the BrowserRouter
- Use custom PropsRoute function that renders components with props in a simpler way
- Navigate through pages through Link components