refactor: Refactor books state to shelves state
Reorganize the books state separated by shelves for better performance,
as we don't have to iterate over the books collection all the time to
find the ones that belongs to a shelf. It also makes it easier to work
with the BooksAPI.update call response format, as the response objects
shows a list of shelves rather a list of books.

- Reduce books list using shelf as key and list of books as the value;